backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.2.0,<6.0 # For JWT authentication
psycopg2-binary>=2.9,<3.0
celery>=5.3,<6.0
redis>=4.0,<5.0
gunicorn>=20.0,<21.0 # For production deployment
django-storages[google]>=1.14,<1.15 # For Google Cloud Storage
pgvector>=0.2.0,<0.3.0 # For vector embeddings in PostgreSQL
django-pgvector>=0.2.0,<0.3.0 # Django integration for pgvector
transformers>=4.30.0,<5.0.0 # For generating embeddings
torch>=2.0.0,<3.0.0 # For PyTorch backend (used by transformers)
spacy>=3.7.0,<4.0.0 # For NLP pipeline
sentence-transformers>=2.2.2,<3.0.0 # For sentence/embedding models
pdf2image>=1.16.3,<2.0.0 # For PDF to image conversion
pillow>=10.0.0,<11.0.0 # For image processing
PyPDF2>=3.0.0,<4.0.0 # For PDF metadata extraction
